Move CharacterCardPico back to PageCharacter
